+++ packages/flutter_image/test/network_test.dart
@@ -142,7 +142,7 @@ void assertThatImageLoadingFails(
 ) {
   final ImageStreamCompleter completer = subject.load(
     subject,
-    _ambiguate(PaintingBinding.instance)!.instantiateImageCodec,
+    PaintingBinding.instance.instantiateImageCodec,
   );
   completer.addListener(ImageStreamListener(
     (ImageInfo image, bool synchronousCall) {},
@@ -159,7 +159,7 @@ void assertThatImageLoadingSucceeds(
 ) {
   final ImageStreamCompleter completer = subject.load(
     subject,
-    _ambiguate(PaintingBinding.instance)!.instantiateImageCodec,
+    PaintingBinding.instance.instantiateImageCodec,
   );
   completer.addListener(ImageStreamListener(
     expectAsync2((ImageInfo image, bool synchronousCall) {
@@ -168,11 +168,3 @@ void assertThatImageLoadingSucceeds(
     }),
   ));
 }
-
-/// This allows a value of type T or T? to be treated as a value of type T?.
-///
-/// We use this so that APIs that have become non-nullable can still be used
-/// with `!` and `?` on the stable branch.
-// TODO(ianh): Remove this once the relevant APIs have shipped to stable.
-// See https://github.com/flutter/flutter/issues/64830
-T? _ambiguate<T>(T? value) => value;