+++ packages/pigeon/lib/generator_tools.dart
@@ -5,10 +5,11 @@
 import 'dart:convert';
 import 'dart:io';
 import 'dart:mirrors';
+
 import 'ast.dart';
 
 /// The current version of pigeon. This must match the version in pubspec.yaml.
-const String pigeonVersion = '3.1.1';
+const String pigeonVersion = '3.1.2';
 
 /// Read all the content from [stdin] to a String.
 String readStdin() {